Move sidebar context to context folder
